# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CHALLENGE: Operacíon Fuego de Quasar",
        "title": "MELI CHALLENGE",
        "contact": {
            "name": "API Support",
            "email": "gabriel.adrian.felipe@gmail.com"
        },
        "license": {
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
        "/topsecret": {
            "post": {
                "description": "Valida posicion de nave en base a array de satelites",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Valida posicion de nave",
                "parameters": [
                    {
                        "description": "Save Satellite Data",
                        "name": "satelliteData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/spaceModels.Satellites"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.ReponseTopsecretFormat"
                        }
                    }
                }
            }
        },
        "/topsecret_split": {
            "get": {
                "description": "Valida posicion de nave en base a cache",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Valida posicion de nave",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.ReponseTopsecretFormat"
                        }
                    }
                }
            },
            "post": {
                "description": "Guarda en cache datos por cada satelite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Guarda en cache datos por cada satelite",
                "parameters": [
                    {
                        "description": "Save Satellite Data",
                        "name": "satelliteData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/spaceModels.Satellite"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "helpers.ReponsePositionFormat": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
        "helpers.ReponseTopsecretFormat": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "position": {
                    "$ref": "#/definitions/helpers.ReponsePositionFormat"
                }
            }
        },
        "spaceModels.Satellite": {
            "type": "object",
            "properties": {
                "coords": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "distance": {
                    "type": "number"
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "spaceModels.Satellites": {
            "type": "object",
            "properties": {
                "satellites": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/spaceModels.Satellite"
                    }
                }
            }
        }
    }
}